requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
pydantic[email]
pymupdf
python-multipart
